+++ CHANGELOG.md
@@ -8,7 +8,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - Added `running_on_lambda()` to support programs running both locally and on AWS Lambda.
  - Added `LocalContext` and `run_local_handler` to support executing message handlers locally.
- - Updates AWS SDK dependencies to:
+ - **Breaking:** Updated `LambdaContext` to require an `EventType` type, which implements the `RunnableEventType` trait. To resolve this in your existing code, change the definition:
+
+```rust
+ impl LambdaContext<Env> for Context {
+```
+to
+
+```rust
+use cobalt_aws::lambda::SqsEvent;
+
+impl LambdaContext<Env, SqsEvent> for Context {
+```
+ - **Breaking:** Updates AWS SDK dependencies to:
     - `aws-config = "0.55.3"`
     - `aws-sdk-athena = "0.28.0"`
     - `aws-sdk-s3 = "0.28.0"`
@@ -17,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - `aws-types = "0.55.3"`
     - `aws_lambda_events = "0.10.0"`
     - `lambda_runtime = "0.8.1"`
+
+
 
 ## 0.10.0